Fix Slackbot autoloading
